refactor(web): rename llmConfig → languageModels to match sidebar tab name
Renames:
- lib/llmConfig/ → lib/languageModels/
- sections/modals/llmConfig/ → sections/modals/languageModels/
- LLMConfigurationPage → LanguageModelsPage

Updates all ~30 import sites and the Next.js route re-export.
