+++ web/src/app/admin/configuration/language-models/ModelIcon.tsx
@@ -1,5 +1,5 @@
 import { defaultTailwindCSS } from "@/components/icons/icons";
-import { getModelIcon } from "@/lib/llmConfig";
+import { getModelIcon } from "@/lib/languageModels";
 import { IconProps } from "@opal/types";
 
 export interface ModelIconProps extends IconProps {
