+++ web/src/app/craft/onboarding/components/BuildOnboardingModal.tsx
@@ -23,8 +23,8 @@ import {
   getDefaultLlmSelection,
 } from "@/app/craft/onboarding/constants";
 import { LLMProviderDescriptor } from "@/interfaces/llm";
-import { LLM_PROVIDERS_ADMIN_URL } from "@/lib/llmConfig/constants";
-import { testApiKeyHelper } from "@/sections/modals/llmConfig/svc";
+import { LLM_PROVIDERS_ADMIN_URL } from "@/lib/languageModels/constants";
+import { testApiKeyHelper } from "@/sections/modals/languageModels/svc";
 import OnboardingInfoPages from "@/app/craft/onboarding/components/OnboardingInfoPages";
 import OnboardingUserInfo from "@/app/craft/onboarding/components/OnboardingUserInfo";
 import OnboardingLlmSetup, {
