+++ web/src/lib/hooks.llmResolver.test.ts
@@ -2,7 +2,7 @@ import {
   getDefaultLlmDescriptor,
   getValidLlmDescriptorForProviders,
 } from "@/lib/hooks";
-import { structureValue } from "@/lib/llmConfig/utils";
+import { structureValue } from "@/lib/languageModels/utils";
 import { LLMProviderDescriptor } from "@/interfaces/llm";
 import { makeProvider } from "@tests/setup/llmProviderTestUtils";
