+++ web/src/refresh-pages/admin/ImageGenerationPage/ImageGenerationContent.tsx
@@ -26,7 +26,7 @@ import { SvgSlash, SvgUnplug } from "@opal/icons";
 import { markdown } from "@opal/utils";
 import { getImageGenForm } from "@/refresh-pages/admin/ImageGenerationPage/forms";
 import ProviderCard from "@/sections/admin/ProviderCard";
-import { getModelIcon } from "@/lib/llmConfig";
+import { getModelIcon } from "@/lib/languageModels";
 
 const NO_DEFAULT_VALUE = "__none__";
