+++ web/src/sections/modals/AgentViewerModal.tsx
@@ -34,7 +34,7 @@ import { formatMmDdYyyy } from "@/lib/dateUtils";
 import { useProjectsContext } from "@/providers/ProjectsContext";
 import { FileCard } from "@/sections/cards/FileCard";
 import DocumentSetCard from "@/sections/cards/DocumentSetCard";
-import { getDisplayName } from "@/lib/llmConfig/utils";
+import { getDisplayName } from "@/lib/languageModels/utils";
 import { useLLMProviders } from "@/hooks/useLanguageModels";
 import { Interactive } from "@opal/core";
