+++ web/src/sections/modals/languageModels/BifrostModal.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { markdown } from "@opal/utils";
 import { useSWRConfig } from "swr";
 import { useFormikContext } from "formik";
 import { InputDivider } from "@opal/layouts";
@@ -8,64 +9,57 @@ import {
   LLMProviderName,
   LLMProviderView,
 } from "@/interfaces/llm";
+import { fetchBifrostModels } from "@/lib/languageModels/svc";
 import {
   useInitialValues,
   buildValidationSchema,
-  BaseLLMFormValues as BaseLLMModalValues,
+  BaseLLMFormValues,
   mergeFetchedModelConfigurations,
-} from "@/sections/modals/llmConfig/utils";
-import { submitProvider } from "@/sections/modals/llmConfig/svc";
+} from "@/sections/modals/languageModels/utils";
+import { submitProvider } from "@/sections/modals/languageModels/svc";
 import { LLMProviderConfiguredSource } from "@/lib/analytics";
 import {
-  APIKeyField,
   APIBaseField,
+  APIKeyField,
   ModelSelectionField,
   DisplayNameField,
   ModelAccessField,
   ModalWrapper,
-} from "@/sections/modals/llmConfig/shared";
-import { fetchModels } from "@/lib/llmConfig/svc";
+} from "@/sections/modals/languageModels/shared";
 import { toast } from "@/hooks/useToast";
-import { refreshLlmProviderCaches } from "@/lib/llmConfig/cache";
+import { refreshLlmProviderCaches } from "@/lib/languageModels/cache";
 
-const DEFAULT_API_BASE = "http://localhost:1234";
-
-interface LMStudioModalValues extends BaseLLMModalValues {
+interface BifrostModalValues extends BaseLLMFormValues {
+  api_key: string;
   api_base: string;
-  custom_config: {
-    LM_STUDIO_API_KEY?: string;
-  };
 }
 
-interface LMStudioModalInternalsProps {
+interface BifrostModalInternalsProps {
   existingLlmProvider: LLMProviderView | undefined;
   isOnboarding: boolean;
 }
 
-function LMStudioModalInternals({
+function BifrostModalInternals({
   existingLlmProvider,
   isOnboarding,
-}: LMStudioModalInternalsProps) {
-  const formikProps = useFormikContext<LMStudioModalValues>();
+}: BifrostModalInternalsProps) {
+  const formikProps = useFormikContext<BifrostModalValues>();
 
   const isFetchDisabled = !formikProps.values.api_base;
 
   const handleFetchModels = async () => {
-    const apiKey = formikProps.values.custom_config?.LM_STUDIO_API_KEY;
-    const initialApiKey = existingLlmProvider?.custom_config?.LM_STUDIO_API_KEY;
-    const data = await fetchModels(LLMProviderName.LM_STUDIO, {
+    const { models, error } = await fetchBifrostModels({
       api_base: formikProps.values.api_base,
-      custom_config: apiKey ? { LM_STUDIO_API_KEY: apiKey } : {},
-      api_key_changed: apiKey !== initialApiKey,
-      name: existingLlmProvider?.name ?? undefined,
+      api_key: formikProps.values.api_key || undefined,
+      provider_name: existingLlmProvider?.name ?? undefined,
     });
-    if (data.error) {
-      throw new Error(data.error);
+    if (error) {
+      throw new Error(error);
     }
     formikProps.setFieldValue(
       "model_configurations",
       mergeFetchedModelConfigurations(
-        data.models,
+        models,
         formikProps.values.model_configurations
       )
     );
@@ -74,14 +68,15 @@ function LMStudioModalInternals({
   return (
     <>
       <APIBaseField
-        subDescription="The base URL for your LM Studio server."
-        placeholder="Your LM Studio API base URL"
+        subDescription="Paste your Bifrost gateway endpoint URL (including API version)."
+        placeholder="https://your-bifrost-gateway.com/v1"
       />
 
       <APIKeyField
-        name="custom_config.LM_STUDIO_API_KEY"
         optional
-        subDescription="Optional API key if your LM Studio server requires authentication."
+        subDescription={markdown(
+          "Paste your API key from [Bifrost](https://docs.getbifrost.ai/overview) to access your models."
+        )}
       />
 
       {!isOnboarding && (
@@ -107,7 +102,7 @@ function LMStudioModalInternals({
   );
 }
 
-export default function LMStudioModal({
+export default function BifrostModal({
   variant = "llm-configuration",
   existingLlmProvider,
   shouldMarkAsDefault,
@@ -119,17 +114,11 @@ export default function LMStudioModal({
 
   const onClose = () => onOpenChange?.(false);
 
-  const initialValues: LMStudioModalValues = {
-    ...useInitialValues(
-      isOnboarding,
-      LLMProviderName.LM_STUDIO,
-      existingLlmProvider
-    ),
-    api_base: existingLlmProvider?.api_base ?? DEFAULT_API_BASE,
-    custom_config: {
-      LM_STUDIO_API_KEY: existingLlmProvider?.custom_config?.LM_STUDIO_API_KEY,
-    },
-  } as LMStudioModalValues;
+  const initialValues: BifrostModalValues = useInitialValues(
+    isOnboarding,
+    LLMProviderName.BIFROST,
+    existingLlmProvider
+  ) as BifrostModalValues;
 
   const validationSchema = buildValidationSchema(isOnboarding, {
     apiBase: true,
@@ -137,30 +126,18 @@ export default function LMStudioModal({
 
   return (
     <ModalWrapper
-      providerName={LLMProviderName.LM_STUDIO}
+      providerName={LLMProviderName.BIFROST}
       llmProvider={existingLlmProvider}
       onClose={onClose}
       initialValues={initialValues}
       validationSchema={validationSchema}
       onSubmit={async (values, { setSubmitting, setStatus }) => {
-        const filteredCustomConfig = Object.fromEntries(
-          Object.entries(values.custom_config || {}).filter(([, v]) => v !== "")
-        );
-
-        const submitValues = {
-          ...values,
-          custom_config:
-            Object.keys(filteredCustomConfig).length > 0
-              ? filteredCustomConfig
-              : undefined,
-        };
-
         await submitProvider({
           analyticsSource: isOnboarding
             ? LLMProviderConfiguredSource.CHAT_ONBOARDING
             : LLMProviderConfiguredSource.ADMIN_PAGE,
-          providerName: LLMProviderName.LM_STUDIO,
-          values: submitValues,
+          providerName: LLMProviderName.BIFROST,
+          values,
           initialValues,
           existingLlmProvider,
           shouldMarkAsDefault,
@@ -182,7 +159,7 @@ export default function LMStudioModal({
         });
       }}
     >
-      <LMStudioModalInternals
+      <BifrostModalInternals
         existingLlmProvider={existingLlmProvider}
         isOnboarding={isOnboarding}
       />
